fix: add configuration property that allows disabling setting of notify-keyspace-events on redis server
Before this fix, when starting encore, spring data redis
 would try to use the CONFIG command to set
notify-keyspace-events on the redis server. If the server did not support the CONFIG command,
which is the case in many managed redis services, encore would fail to start.

With this fix, setting the property redis.keyspace.disable-config-notifications
to true string will stop spring data redis from trying to change the config.
If the property is not set, or set to true, spring data redis will behave as before,
that is use the CONFIG command to set 'notify-keyspace-events'.

Note that this fix does not seem to work when native compile is used.

Signed-off-by: Gustav Grusell <gustav.grusell@eyevinn.se>