feat: support s3 urls for input and output
This commit add support for using s3 urls on the format
s3://<BUCKET>/<KEY> in both input and output.

If ans s3 URL is used as input, a presigned URL is created
and used as input to ffmpeg. The duration of the presigned URLs can be
controlled with the 'remote-files.s3.presignDurationSeconds' config
property.

If an s3 URL is used for 'outputFolder', output will first be stored
locally and then uploaded to s3 once transcoding is finished.

Aws credentials are read with DefaultCredentialsProvider, meaning
aws credentials can be provided in a number of ways, see
https://sdk.amazonaws.com/java/api/latest/software/amazon/awssdk/auth/credentials/DefaultCredentialsProvider.html;

Not that when using s3 urls for input, the presigned URLs will be
shown in the logs. If this is not desirable, setting
logging.config (or env variable LOGGING_CONFIG) to
'classpath:logback-json-mask-s3-presign.xml'
will use a log config that masks the presign query parameters.

By setting env variable REMOTEFILES_S3_ANONYMOUSACCESS to true, s3 urls will be
accessed in anonymous mode, corresponding to using the '--no-sign-request' flag
with the aws cli. Any s3 access key or secrets key configured will be
ignored. Multipart upload will be disabled in this case since the s3
sdk does not support multipart upload when using anonymous access.

Signed-off-by: Gustav Grusell <gustav.grusell@eyevinn.se>