feat: add support for encoding audio separately without segmenting when using segmented encoding
This feature can be enabled for a job by setting 'segmentedEncodingEnabledForAudio' to 'false' for the job.
If this property is not set for a job, the default value from the configuration property
'encore-settings.encoding.segmentedEncoding.enabledForAudio' will be used.
This property defaults to false, which means the default behaviour is to use segemented encoding also
for audio.

Refs: #41

Signed-off-by: Gustav Grusell <gustav.grusell@eyevinn.se>